Moved logic to query_actor.h
