+++ ydb/core/kqp/proxy_service/kqp_script_executions.cpp
@@ -77,96 +77,6 @@ public:
     {}
 };
 
-template<typename TQueryActor, typename TResponse, typename ...TArgs>
-class TQueryRetryActor : public TActorBootstrapped<TQueryRetryActor<TQueryActor, TResponse, TArgs...>> {
-public:
-    using TBase = TActorBootstrapped<TQueryRetryActor<TQueryActor, TResponse, TArgs...>>;
-    using IRetryPolicy = IRetryPolicy<Ydb::StatusIds::StatusCode>;
-
-    explicit TQueryRetryActor(const TActorId& replyActorId, const TArgs&... args, TDuration maxRetryTime = TDuration::Seconds(30))
-        : ReplyActorId(replyActorId)
-        , CreateQueryActor([=]() {
-            return new TQueryActor(args...);
-        })
-        , MaxRetryTime(maxRetryTime)
-    {}
-
-    void StartQueryActor() const {
-        TBase::Register(CreateQueryActor());
-    }
-
-    void Bootstrap() {
-        TBase::Become(&TQueryRetryActor::StateFunc);
-        StartQueryActor();
-    }
-
-    STRICT_STFUNC(StateFunc,
-        hFunc(TEvents::TEvWakeup, Wakeup);
-        hFunc(TResponse, Handle);
-    )
-
-    void Wakeup(TEvents::TEvWakeup::TPtr&) {
-        StartQueryActor();
-    }
-
-    void Handle(const typename TResponse::TPtr& ev) {
-        const Ydb::StatusIds::StatusCode status = ev->Get()->Status;
-        if (Retryable(status) == ERetryErrorClass::NoRetry) {
-            Reply(ev);
-            return;
-        }
-
-        if (RetryState == nullptr) {
-            CreateRetryState();
-        }
-
-        if (auto delay = RetryState->GetNextRetryDelay(status)) {
-            TBase::Schedule(*delay, new TEvents::TEvWakeup());
-        } else {
-            Reply(ev);
-        }
-    }
-
-    void Reply(const typename TResponse::TPtr& ev) {
-        TBase::Send(ev->Forward(ReplyActorId));
-        TBase::PassAway();
-    }
-
-    static ERetryErrorClass Retryable(Ydb::StatusIds::StatusCode status) {
-        if (status == Ydb::StatusIds::SUCCESS) {
-            return ERetryErrorClass::NoRetry;
-        }
-
-        if (status == Ydb::StatusIds::INTERNAL_ERROR
-            || status == Ydb::StatusIds::UNAVAILABLE
-            || status == Ydb::StatusIds::BAD_SESSION
-            || status == Ydb::StatusIds::SESSION_EXPIRED
-            || status == Ydb::StatusIds::SESSION_BUSY
-            || status == Ydb::StatusIds::TIMEOUT
-            || status == Ydb::StatusIds::ABORTED) {
-            return ERetryErrorClass::ShortRetry;
-        }
-
-        if (status == Ydb::StatusIds::OVERLOADED
-            || status == Ydb::StatusIds::UNDETERMINED) {
-            return ERetryErrorClass::LongRetry;
-        }
-
-        return ERetryErrorClass::NoRetry;
-    }
-
-    void CreateRetryState() {
-        IRetryPolicy::TPtr policy = IRetryPolicy::GetExponentialBackoffPolicy(Retryable, TDuration::MilliSeconds(10), TDuration::MilliSeconds(200), TDuration::Seconds(1), std::numeric_limits<size_t>::max(), MaxRetryTime);
-        RetryState = policy->CreateRetryState();
-    }
-
-private:
-    const TActorId ReplyActorId;
-    const std::function<TQueryActor*()> CreateQueryActor;
-    const TDuration MaxRetryTime;
-    IRetryPolicy::IRetryState::TPtr RetryState = nullptr;
-};
-
 
 class TScriptExecutionsTablesCreator : public TActorBootstrapped<TScriptExecutionsTablesCreator> {
 public:
