+++ ydb/library/query_actor/query_actor.h
@@ -12,11 +12,12 @@
 #include <ydb/library/actors/core/actorsystem.h>
 #include <ydb/library/actors/core/event_local.h>
 #include <ydb/library/actors/core/events.h>
+#include <ydb/library/actors/core/hfunc.h>
+#include <library/cpp/retry/retry_policy.h>
 #include <library/cpp/threading/future/future.h>
 
 namespace NKikimr {
 
-// TODO: add retry logic
 class TQueryBase : public NActors::TActorBootstrapped<TQueryBase> {
 protected:
     struct TTxControl {
@@ -166,6 +167,103 @@ protected:
     NActors::TActorId Owner;
 
     std::vector<NYdb::TResultSet> ResultSets;
+};
+
+template<typename TQueryActor, typename TResponse, typename ...TArgs>
+class TQueryRetryActor : public NActors::TActorBootstrapped<TQueryRetryActor<TQueryActor, TResponse, TArgs...>> {
+public:
+    using TBase = NActors::TActorBootstrapped<TQueryRetryActor<TQueryActor, TResponse, TArgs...>>;
+    using IRetryPolicy = IRetryPolicy<Ydb::StatusIds::StatusCode>;
+
+    explicit TQueryRetryActor(const NActors::TActorId& replyActorId, const TArgs&... args, TDuration maxRetryTime = TDuration::Seconds(1))
+        : ReplyActorId(replyActorId)
+        , RetryPolicy(IRetryPolicy::GetExponentialBackoffPolicy(
+            Retryable, TDuration::MilliSeconds(10), 
+            TDuration::MilliSeconds(200), TDuration::Seconds(1),
+            std::numeric_limits<size_t>::max(), maxRetryTime
+        ))
+        , CreateQueryActor([=]() {
+            return new TQueryActor(args...);
+        })
+    {}
+
+    TQueryRetryActor(const NActors::TActorId& replyActorId, IRetryPolicy::TPtr retryPolicy, const TArgs&... args)
+        : ReplyActorId(replyActorId)
+        , RetryPolicy(retryPolicy)
+        , CreateQueryActor([=]() {
+            return new TQueryActor(args...);
+        })
+    {}
+
+    void StartQueryActor() const {
+        TBase::Register(CreateQueryActor());
+    }
+
+    void Bootstrap() {
+        TBase::Become(&TQueryRetryActor::StateFunc);
+        StartQueryActor();
+    }
+
+    STRICT_STFUNC(StateFunc,
+        hFunc(NActors::TEvents::TEvWakeup, Wakeup);
+        hFunc(TResponse, Handle);
+    )
+
+    void Wakeup(NActors::TEvents::TEvWakeup::TPtr&) {
+        StartQueryActor();
+    }
+
+    void Handle(const typename TResponse::TPtr& ev) {
+        const Ydb::StatusIds::StatusCode status = ev->Get()->Status;
+        if (Retryable(status) == ERetryErrorClass::NoRetry) {
+            Reply(ev);
+            return;
+        }
+
+        if (RetryState == nullptr) {
+            RetryState = RetryPolicy->CreateRetryState();
+        }
+
+        if (auto delay = RetryState->GetNextRetryDelay(status)) {
+            TBase::Schedule(*delay, new NActors::TEvents::TEvWakeup());
+        } else {
+            Reply(ev);
+        }
+    }
+
+    void Reply(const typename TResponse::TPtr& ev) {
+        TBase::Send(ev->Forward(ReplyActorId));
+        TBase::PassAway();
+    }
+
+    static ERetryErrorClass Retryable(Ydb::StatusIds::StatusCode status) {
+        if (status == Ydb::StatusIds::SUCCESS) {
+            return ERetryErrorClass::NoRetry;
+        }
+
+        if (status == Ydb::StatusIds::INTERNAL_ERROR
+            || status == Ydb::StatusIds::UNAVAILABLE
+            || status == Ydb::StatusIds::BAD_SESSION
+            || status == Ydb::StatusIds::SESSION_EXPIRED
+            || status == Ydb::StatusIds::SESSION_BUSY
+            || status == Ydb::StatusIds::TIMEOUT
+            || status == Ydb::StatusIds::ABORTED) {
+            return ERetryErrorClass::ShortRetry;
+        }
+
+        if (status == Ydb::StatusIds::OVERLOADED
+            || status == Ydb::StatusIds::UNDETERMINED) {
+            return ERetryErrorClass::LongRetry;
+        }
+
+        return ERetryErrorClass::NoRetry;
+    }
+
+private:
+    const NActors::TActorId ReplyActorId;
+    const IRetryPolicy::TPtr RetryPolicy;
+    const std::function<TQueryActor*()> CreateQueryActor;
+    IRetryPolicy::IRetryState::TPtr RetryState = nullptr;
 };
 
 } // namespace NKikimr
